% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{as_df}
\alias{as_df}
\alias{as_df,methCall-method}
\title{Convert methCall object to data frame}
\usage{
as_df(x, name)

\S4method{as_df}{methCall}(x)
}
\arguments{
\item{x}{A methCall object}
}
\description{
Convert methCall object to data frame
}
