% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{as_GRanges}
\alias{as_GRanges}
\alias{as_GRanges,methCall-method}
\title{Convert methCall object to data frame}
\usage{
as_GRanges(x)

\S4method{as_GRanges}{methCall}(x)
}
\arguments{
\item{x}{A methCall object}
}
\description{
Convert methCall object to data frame
}
