% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvd.R
\name{choose_k}
\alias{choose_k}
\title{Decide the rank of the truncated SVD based on Linderman 2018 bioRxiv}
\usage{
choose_k(A, K = 100, pval_thresh = 1e-10, noise_start = 80, q = 2)
}
\arguments{
\item{A}{array_like; \cr
a real/complex \eqn{(m, n)} input matrix (or data frame) to be decomposed.}

\item{pval_thresh}{The threshold for ``significance''}

\item{noise_start}{Index for which all smaller singular values are considered noise}

\item{q}{integer, optional; \cr
number of additional power iterations (by default \eqn{q=2}).}

\item{n}{Number of resampling iteration}
}
\value{
A list with three items: Chosen k, P values of each possible k, 
singular values of the matrix A
}
\description{
Adpated from https://github.com/KlugerLab/ALRA/blob/master/alra.R
Heuristic for choosing rank k for the low rank approximation based on
statistics of the spacings between consecutive singular values. Finds
the smallest singular value $\sigma_i$ such that $\sigma_i - \sigma_{i-1}$
is significantly different than spacings in the tail of the singular values.
}
