% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{extract_methylation}
\alias{extract_methylation}
\title{Extract the relevant columns from the files}
\usage{
extract_methylation(dir, col_idx, out_dir)
}
\arguments{
\item{dir}{The directory containing tsv.gz files for methylation status}

\item{col_idx}{The index of column for chromosome, start, methylation level}

\item{out_dir}{The output directory}
}
\description{
Extract the relevant columns from the files
}
