% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{filter_sites}
\alias{filter_sites}
\alias{filter_sites,methCall,function-method}
\title{Filter sites}
\usage{
filter_sites(x, f)

\S4method{filter_sites}{methCall,`function`}(x, f)
}
\arguments{
\item{x}{A methCall object}

\item{f}{A predicate function that return a logical vector for whether to keep 
the sites}
}
\description{
Filter sites
}
