% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{join_meth_list}
\alias{join_meth_list}
\title{Join a list of methCall object by site}
\usage{
join_meth_list(meth_list, all.x = FALSE, all.y = FALSE)
}
\arguments{
\item{meth_list}{A named list of methCall object (from read_meth)}

\item{all.x}{Whether to include all rows in x}

\item{all.y}{Whether to include all rows in y}
}
\value{
A DataFrame that is a inner join of all the methCall
}
\description{
Join a list of methCall object by site
}
