% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{map_coord}
\alias{map_coord}
\alias{map_coord,methCall,character-method}
\title{Liftover methCall object coordinate}
\usage{
map_coord(x, y)

\S4method{map_coord}{methCall,character}(x, y)
}
\arguments{
\item{x}{A methCall object}

\item{y}{The path to the chain file}
}
\description{
Liftover methCall object coordinate
}
