% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{methCall-class}
\alias{methCall-class}
\title{An S4 class to store methylation call object}
\description{
An S4 class to store methylation call object
}
\section{Slots}{

\describe{
\item{\code{data}}{A DataFrame object containing 6 columns (chr, position, strand,
met_reads, unmet_reads)}
}}

