% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{n_row}
\alias{n_row}
\alias{n_row,methCall-method}
\title{Number of rows}
\usage{
n_row(x)

\S4method{n_row}{methCall}(x)
}
\arguments{
\item{x}{A methCall object}
}
\description{
Number of rows
}
