% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock.R
\name{PhenoAge}
\alias{PhenoAge}
\title{Compute DNAm PhenoAge}
\usage{
PhenoAge(data)
}
\arguments{
\item{data}{A matrix with rows as CpGs and columns as samples (rownames should
be probe id)}
}
\value{
PhenoAge prediction
}
\description{
Compute DNAm PhenoAge
}
