% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot coverage distribution for a list of methCall objects}
\usage{
plot_coverage(meth_list)
}
\arguments{
\item{meth_list}{A list of methCall objects}
}
\value{
A list of ggplot objects
}
\description{
Plot coverage distribution for a list of methCall objects
}
