% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{read_meth}
\alias{read_meth}
\title{Extract the relevant columns from the files}
\usage{
read_meth(files, chr_idx, pos_idx, met_idx, id, coverage_idx = NULL,
  strand_idx = NULL, unmet_idx = NULL, deduplicate = T, header = T,
  sep = "\\t")
}
\arguments{
\item{files}{The path to tsv files for methylation status}

\item{chr_idx}{The index of column for chromosome}

\item{pos_idx}{The index of column for position}

\item{met_idx}{The index of column for the number of methylated reads}

\item{id}{The id to each file}

\item{coverage_idx}{The index of column containing coverage; optional}

\item{strand_idx}{The index of column containing strand; optional}

\item{unmet_idx}{The index of column for the number of unmethylated reads}

\item{deduplicate}{Whether to sum the reads for rows with the same genomic location}

\item{header}{Whether to there is a header}

\item{sep}{delimiter for methylation status file}
}
\description{
Extract the relevant columns from the files
}
