% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvd.R
\name{SVT_denoise}
\alias{SVT_denoise}
\title{Denoise using truncated SVD}
\usage{
SVT_denoise(A, lambda, svd = NULL)
}
\arguments{
\item{A}{array_like; \cr
a real/complex \eqn{(m, n)} input matrix (or data frame) to be decomposed.}
}
\value{
A matrix of denoised data
}
\description{
Denoise using truncated SVD
}
