% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_coef}
\alias{tidy_coef}
\title{Get coefficients from glmnet}
\usage{
tidy_coef(x, s, coef_names = NULL)
}
\arguments{
\item{x}{A glmnet object}

\item{s}{lambda value for glmnet}

\item{coef_names}{The names of the coefficients}
}
\description{
Get coefficients from glmnet
}
