% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvd.R
\name{TSVD_denoise}
\alias{TSVD_denoise}
\title{Denoise using truncated SVD}
\usage{
TSVD_denoise(A, k = 30, ...)
}
\arguments{
\item{A}{array_like; \cr
a real/complex \eqn{(m, n)} input matrix (or data frame) to be decomposed.}

\item{k}{integer; \cr
specifies the target rank of the low-rank decomposition. \eqn{k} should satisfy \eqn{k << min(m,n)}.}
}
\description{
Denoise using truncated SVD
}
